// src/enums/opcode-branch.ts
export enum OpcodeBranch {
    ZERO = 0x0000,
    GOTO = 0x0001,
    CALL_SUBROUTINE = 0x0002,
    SKIP_NEXT_IF_EQUALS = 0x0003,
    SKIP_NEXT_IF_NOT_EQUALS = 0x0004,
    SKIP_NEXT_IF_REGISTER_EQUAL = 0x0005,
    SET_REGISTER_TO_VAL = 0x0006,
    ADD_TO_REGISTER = 0x0007,
    REGISTER_OP = 0x0008
}

export enum RegisterOperation {
    SET_REGISTER_TO_REGISTER = 0x0000,
    SET_REGISTER_OR = 0x0001,
    SET_REGISTER_AND = 0x0002,
    SET_REGISTER_XOR = 0x0003,
    ADD_REGISTER = 0x0004,
    SUBSTRACT_REGISTER = 0x0005,
    SET_REGISTER_SHIFT_RIGHT = 0x0006,
    SUBSTRACT_REGISTER_OPPOSITE = 0x0007,
    SET_REGISTER_SHIFT_LEFT = 0x000e,
}